Update on the model and Repository files with use of Lomok Addedd StudentApplication Model, repo and service and some corrections Use Long Instead of int for id (i.e immutable type)

+++ backend/demo/demo/src/main/java/testing/demo/Services/UserCredentialsService.java
@@ -20,14 +20,14 @@ public class UserCredentialsService {
        return repo.save(userCredentials)  ;
     }
 
-    public UserCredentials findPersonById(int id){
+    public UserCredentials findPersonById(Long id){
        return repo.findById(id).orElse(null);
     }
     public List<UserCredentials> getAllUser(){
         return  repo.findAll();
     }
 
-    public String deleteById(int id){
+    public String deleteById(Long id){
         repo.deleteById(id);
         return "A person with  Id  " +id +" is deleted!!";
     }
